+++ razorboard/src/app/http-loading.interceptor.ts
@@ -2,11 +2,11 @@ import { Injectable } from '@angular/core';
 import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { finalize } from 'rxjs/operators';
-import { HttpLoadingService } from './http-loading.service';
+import { HttpEventsService } from './http-events.service';
 
 @Injectable()
 export class HttpLoadingInterceptor implements HttpInterceptor {
-  constructor(public loaderService: HttpLoadingService) { }
+  constructor(public loaderService: HttpEventsService) { }
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     this.loaderService.show();
     return next.handle(req).pipe(