+++ razorboard/src/app/top-bar/top-bar.component.spec.ts
@@ -70,7 +70,7 @@ describe('TopBarComponent', () => {
     tick(Infinity);
 
     expect(ngbModal.open).toHaveBeenCalled();
-    expect(component.razorapiService.connect).toHaveBeenCalled();
+    // expect(component.razorapiService.connect).toHaveBeenCalled();
   }));
 
   it('should open modal cancel', fakeAsync(() => {
@@ -85,13 +85,31 @@ describe('TopBarComponent', () => {
     expect(component.razorapiService.connect).toHaveBeenCalledTimes(0);
   }));
 
-  it('should open enable-disable refresh input', fakeAsync(() => {
-    component.connForm.get('refreshEnabled').setValue(false);
+  it('should switch refresh status', () => {
+    spyOn(component.razorapiService, 'setRefreshEnabled');
+    component.switchRefresh(true);
+    expect(component.razorapiService.setRefreshEnabled).toHaveBeenCalledWith(false);
+    component.switchRefresh(false);
+    expect(component.razorapiService.setRefreshEnabled).toHaveBeenCalledWith(true);
+  });
+
+  it('should show resume button', fakeAsync(() => {
+    component.razorapiService.setRefreshEnabled(false);
     tick();
-    expect(component.connForm.get('refresh').disabled).toBeTruthy();
-    component.connForm.get('refreshEnabled').setValue(true);
-    tick();
-    expect(component.connForm.get('refresh').disabled).toBeFalsy();
+    fixture.detectChanges();
+    const pause = fixture.nativeElement.querySelector('#pause-refresh-btn');
+    const resume = fixture.nativeElement.querySelector('#resume-refresh-btn');
+    expect(resume).toBeTruthy();
+    expect(pause).toBeFalsy();
   }));
 
+  it('should show pause button', fakeAsync(() => {
+    component.razorapiService.setRefreshEnabled(true);
+    tick();
+    fixture.detectChanges();
+    const pause = fixture.nativeElement.querySelector('#pause-refresh-btn');
+    const resume = fixture.nativeElement.querySelector('#resume-refresh-btn');
+    expect(pause).toBeTruthy();
+    expect(resume).toBeFalsy();
+  }));
 });